+++ common/src/main/java/net/dawson/adorablehamsterpets/entity/custom/HamsterEntity.java
@@ -151,6 +151,8 @@ public class HamsterEntity extends TameableEntity implements GeoEntity, Implemen
             DataTracker.registerData(HamsterEntity.class, TrackedDataHandlerRegistry.ITEM_STACK);
     private static final TrackedData<ItemStack> TRACKED_ARMOR_STACK =
             DataTracker.registerData(HamsterEntity.class, TrackedDataHandlerRegistry.ITEM_STACK);
+    private static final TrackedData<Boolean> ARMOR_VISIBLE =
+            DataTracker.registerData(HamsterEntity.class, TrackedDataHandlerRegistry.BOOLEAN);
     private static final TrackedData<Boolean> FALL_IMMUNITY_ACTIVE =
             DataTracker.registerData(HamsterEntity.class, TrackedDataHandlerRegistry.BOOLEAN);
     private static final TrackedData<String> ACTIVE_CUSTOM_GOAL_NAME_DEBUG =
@@ -325,6 +327,7 @@ public class HamsterEntity extends TameableEntity implements GeoEntity, Implemen
                 SHOULDER_ANIMATION_STATE, ShoulderAnimationState.STANDING.ordinal());
         this.dataTracker.startTracking(TRACKED_ACCESSORY_STACK, ItemStack.EMPTY);
         this.dataTracker.startTracking(TRACKED_ARMOR_STACK, ItemStack.EMPTY);
+        this.dataTracker.startTracking(ARMOR_VISIBLE, true);
         this.dataTracker.startTracking(FALL_IMMUNITY_ACTIVE, true);
     }
 
@@ -606,6 +609,14 @@ public class HamsterEntity extends TameableEntity implements GeoEntity, Implemen
 
     public void setCheekPouchUnlocked(boolean unlocked) {
         setHamsterFlag(CHEEK_POUCH_UNLOCKED_FLAG, unlocked);
+    }
+
+    public boolean isArmorVisible() {
+        return this.dataTracker.get(ARMOR_VISIBLE);
+    }
+
+    public void setArmorVisible(boolean visible) {
+        this.dataTracker.set(ARMOR_VISIBLE, visible);
     }
 
     public HamsterGenome getGenome() {